requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.15
pandas==2.0.2
numpy==1.24.3
pymysql==1.1.0
openpyxl==3.1.2
python-docx==0.8.11
matplotlib==3.7.1
seaborn==0.12.2
schedule==1.2.0
python-dotenv==1.0.0